op: Remove totp tools since op supports those now
I think that was added in 0.5.4 or something, but the important thing
is I don't need to keep around the awful code for extracting an OTP URI
from op.

Incidentally, there is an uncommitted change to this that added support
for oathtool until I realized I could just get the totp from op.